Refactor filenames and fix linting warnings
