+++ spec/dummy/client/renderer.js
@@ -7,9 +7,8 @@ const path = require('path');
 const reactOnRailsRenderer = require('../../../node_package/src/ReactOnRailsRenderer');
 
 const config = {
-  bundlePath: path.resolve(__dirname, '../tmp'),  // Save bundle to "tmp/" dir of our dummy app
-  port: 3800,                                     // Listen at port 3800
-  vm: 'sandbox',                                  // Switch to 'sandbox' if you have TypeError on Proxy
+  bundlePath: path.resolve(__dirname, '../tmp/bundles'),  // Save bundle to "tmp/" dir of our dummy app
+  port: 3800,                                             // Listen at port 3800
 };
 
 reactOnRailsRenderer(config);
